Fix use of deprecated datetime.utcnow() function
This function has been deprecated in Python 3.12, as seen in this warning:
```
  /usr/local/lib/python3.12/site-packages/appstoreserverlibrary/api\_client.py:469: DeprecationWarning: datetime.datetime.utcnow() is deprecated and scheduled for removal in a future version. Use timezone-aware objects to represent datetimes in UTC: datetime.datetime.now(datetime.UTC).
    future\_time = datetime.datetime.utcnow() + datetime.timedelta(minutes=5)
```

You can see that warning when running the tests in this repository with a sufficiently new Python (`python3 -m unittest`). Luckily the replacement has been available for a while! I've used datetime.timezone.utc instead of datetime.UTC because that alias was only introduced in 3.11.